requirements.txt
# =============================================================================
# Databricks PS Risk Assessment Tool - Python Dependencies
# =============================================================================

# -----------------------------------------------------------------------------
# Core Backend Framework
# -----------------------------------------------------------------------------
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0

# -----------------------------------------------------------------------------
# Databricks Integration
# -----------------------------------------------------------------------------
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0

# -----------------------------------------------------------------------------
# AI / Machine Learning (Hugging Face)
# -----------------------------------------------------------------------------
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
sentencepiece>=0.1.99

# -----------------------------------------------------------------------------
# Dash Dashboards
# -----------------------------------------------------------------------------
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=1.5.0

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
numpy>=1.24.0
pyarrow>=14.0.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
requests>=2.31.0
pydantic>=2.5.0
uuid>=1.30

# -----------------------------------------------------------------------------
# Development & Testing
# -----------------------------------------------------------------------------
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
